From Mathias Froehlich, "attached is a change to the ProxyNode that also includes ProxyNode local database options like recently added to PagedLOD.
Also there is a change to the traverse method:
The previous ProxyNode checks the VisitorType to be a CULL_VISITOR and the
presence of a request handler to submit a database request.
In contrast to that PagedLOD uses the request handler if it is there - even if
the visitor type is not a cull visitor.
The change removes the cull visitor test from the ProxyNode so that it behaves
like the PagedLOD.
I believe that the presence of a request handler in a visitor might be
sufficient to trigger the requests as this is done in the PagedLOD anyway.

Based on rev 10332."


git-svn-id: http://svn.openscenegraph.org/osg/OpenSceneGraph/trunk@10341 16af8721-9629-0410-8352-f15c8da7e697