From Alberto Luaces, "the new changes in osgAnimation showed some errors on the Cygwin platform:
1 - osgAnimation/StatsHandler includes <cmath>, but on Cygwin only std::isnan
is defined that way. I changed it to <math.h> as it is done in the rest of
the OSG.

2 - Pulling some osgViewer headers from the same file in osgAnimation makes
the former a new dependency. Thus, I got errors at osgAnimation linking time.
I wonder why nobody noticed this on other platforms. I have updated the
CMakeLists.txt file in order to link to the needed libraries, and

3 - I have updated the openscenegraph-osgAnimation.pc.in file accordingly."


git-svn-id: http://svn.openscenegraph.org/osg/OpenSceneGraph/trunk@10352 16af8721-9629-0410-8352-f15c8da7e697