From Terry Welsh, "Okay, here are the mods I wrote for drawing a filled bounding box behind osgText::Text.  I made it so the box would get drawn using whichever BackdropImplementation was selected.  However, I did not implement STENCIL_BUFFER.  In that case it defaults to drawing the bounding box using POLYGON_OFFSET instead.
Also made it so the BOUNDINGBOX and FILLEDBOUNDINGBOX are drawn with a
settable color and margin size.

While I was at it I tightened up the values applied with DEPTH_RANGE
and POLYGON_OFFSET, not just for drawing the bounding box but also for
drawing backdrop text (these values must be coupled since the bounding
box has to be drawn deeper in Z than the backdrop text).  The values
in use before seemed like overkill and I was seeing some z-clipping
with my background scenery in the case of DEPTH_RANGE.  If there was a
good reason for the large values please let me know...."


git-svn-id: http://svn.openscenegraph.org/osg/OpenSceneGraph/trunk@10505 16af8721-9629-0410-8352-f15c8da7e697