For testing purposes added code path and options for testing viewer creation in series, including with enabling of VBO's. Options are:
 osgcamera -r 5 --vbo cow.osg

Which repeats construction of the viewer 5 times in a row, and enables VBO, and on each repeat a new model is loaded.

  osgcamera -r 2 --vbo --shared cow.osg

Which repeats construction of the viewer 2 times in a row, and enables VBO, and on each loads the model once and shares it between each instance of the viewer.



git-svn-id: http://svn.openscenegraph.org/osg/OpenSceneGraph/trunk@10831 16af8721-9629-0410-8352-f15c8da7e697