From Sukender, "I've fixed positions for 3DS writer. Points in 3DS must be in world coordinates and I added what was missing. And by refactoring a bit of code, I may have fixed some StateSets related bugs (was ignoring StateSets for osg::Groups). I also added support for Billboard's points, so now "osgconv lz.osg lz.3ds" has an acceptable output. However, there is no rotation depending on billboards' axis, hence the notice "Warning: 3DS writer is incomplete for Billboards (rotation not implemented).". You may want to remove this notice (or lower the notify severity) if you feel 3DS doesn't have to handle such rotations. The attached archive contains 3 files from 3DS plugin, against rev. 11162.
Please note there is still the textures issue for cow.osg. I guess it's because it's not a "flat, dummy and standard" texture in slot 0... That is to say the only thing the writer can handle at the moment. I guess I won't address this soon.
"
and

"I've detected and fixed another bug in 3DS writer: support for automatic splitting of meshes having >65k faces/points was buggy (was deleting faces).
Here is my four 3DS modified files (in a ZIP), against rev. 11193, including previous fixes AND Stephan's fix about relative filenames."



git-svn-id: http://svn.openscenegraph.org/osg/OpenSceneGraph/trunk@11194 16af8721-9629-0410-8352-f15c8da7e697