From Stephan Huber and Mathieu Marache, "attached you'll find framework support for os x via cmake. Please credit Mathieu Marache, he added the last missing piece to this puzzle.
I think it is safe to commit these changes to trunk, as the traditional
way via dylibs should work as before.

Here's some more info how to get frameworks:

With these modifications it is possible to compile frameworks on OS X,
when you set the Cmake-option OSG_COMPILE_FRAMEWORKS to true. If you
want to embed the frameworks in your app-bundle make sure to set
OSG_COMPILE_FRAMEWORKS_INSTALL_NAME_DIR accordingly.

You'll have to build the install-target of the generated xcode-projects
as this sets the install_name_dirs of the frameworks and plugins."


git-svn-id: http://svn.openscenegraph.org/osg/OpenSceneGraph/trunk@11341 16af8721-9629-0410-8352-f15c8da7e697