From Colin McDonald, "I have an application with multiple windows.  They share GL objects between the contexts, using the GraphicsContext::Traits sharedContext and setting the same contextID.
When one of these shared contexts is closed, GraphicsContext::close
deletes all GLObjects for that contextID, regardless of the fact that
they are shared.  This means that all of the other contexts sharing the
objects have to recompile them.

The attached tweak makes GraphicsContext::close a bit less brutal for
shared contexts.  I have also changed a misleading diagnostic message.
"


git-svn-id: http://svn.openscenegraph.org/osg/OpenSceneGraph/trunk@12935 16af8721-9629-0410-8352-f15c8da7e697