From Colin McDonald, "Upgrading to OpenSceneGraph 3.0.1, texture mip mapping stopped working on some junk low-end graphics cards which I still have to support.  They worked ok with osg 2.8 and earlier.
The problem turned out to be with gl proxy textures, which are
unreliable on those devices.  Proxy textures are used by the glu
mipmap build routines to determine if a texture size is
supported. The external glu library had a nice fallback
behaviour, so that if proxy textures didn't work then the mipmap
texture was still created.  But in the work on the new embedded
glu routines that fallback behaviour has been inadvertently
lost.  I have restored the fallback in
src/osg/glu/libutil/mipmap.cpp.  It doesn't add any extra
complexity."


git-svn-id: http://svn.openscenegraph.org/osg/OpenSceneGraph/trunk@12981 16af8721-9629-0410-8352-f15c8da7e697